[Finishes #186464578] Add grocery list removal and tweak UI

+++ backend/pantryVision-core/src/main/java/com/pantryvisioncore/repository/GroceryListRepository.java
@@ -8,5 +8,5 @@ import com.pantryvisioncore.model.GroceryList;
 
 public interface GroceryListRepository extends JpaRepository<GroceryList, Long> {
     List<GroceryList> findByUserId(long userId);
-    List<GroceryList> findByGroceryListId(int listId);
+    List<GroceryList> findByGroceryListId(long listId);
 }
